Rename directories and move files around a little
