+++ src/page/home/MainView.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {View} from 'react-native';
 import PropTypes from 'prop-types';
 import _ from 'underscore';
-import ReportView from './Report/ReportView';
+import ReportView from './report/ReportView';
 import WithIon from '../../components/WithIon';
 import IONKEYS from '../../IONKEYS';
 import styles from '../../style/StyleSheet';
